Edited all these classes, to correctly add the cupcake to the basket, and save it to the database

+++ src/main/java/app/entities/Order.java
@@ -6,10 +6,10 @@ public class Order {
     private int orderId;
     private int userId;
     private List<ProductLine> productList;
-    private Double orderPrice;
+    private float orderPrice;
     private boolean isPaid;
 
-    public Order(int orderId, int userId, Double orderPrice) {
+    public Order(int orderId, int userId, float orderPrice) {
         this.orderId = orderId;
         this.userId = userId;
         this.orderPrice = orderPrice;
@@ -23,7 +23,7 @@ public class Order {
         return userId;
     }
 
-    public Double getOrderPrice() {
+    public float getOrderPrice() {
         return orderPrice;
     }
 }
